feat: Got table rendering rows, columns and actions
